fix PageIndexServiceImpl and SiteManagerImpl in part of multithreading access

+++ src/main/java/ru/belkov/SiteSearchEngine/model/SiteManagerImpl.java
@@ -11,7 +11,7 @@ public class SiteManagerImpl implements SiteManager {
     private PageIndexService pageIndexService;
     private Site site;
     private Thread thread;
-    private boolean stop = true;
+    private volatile boolean stop = true;
     private static final Logger logger = LoggerFactory.getLogger(SiteManagerImpl.class);
 
     public SiteManagerImpl(Site site, SiteService siteService, PageIndexService pageIndexService) {
@@ -21,23 +21,26 @@ public class SiteManagerImpl implements SiteManager {
     }
 
     @Override
-    public void startParsing() {
-        if (site != null) {
-            try {
-                if (thread != null && thread.isAlive()) {
-                    thread.join();
-                }
-            } catch (Exception e) {
-                logger.error(e.toString());
-            }
+    public synchronized void startParsing() {
+        if (stop) {
             stop = false;
-            siteService.deleteSiteByUrl(site);
-            site.setLastError("");
-            site.setStatus(SiteStatus.INDEXING);
-            site = siteService.addIfNotExists(site);
-            SiteThread siteThread = new SiteThread(site, this, pageIndexService, siteService);
-            thread = new Thread(siteThread);
-            thread.start();
+            if (site != null) {
+                try {
+                    if (thread != null && thread.isAlive()) {
+                        thread.join();
+                    }
+                } catch (Exception e) {
+                    logger.error(e.toString());
+                }
+                stop = false;
+                siteService.deleteSiteByUrl(site);
+                site.setLastError("");
+                site.setStatus(SiteStatus.INDEXING);
+                site = siteService.addIfNotExists(site);
+                SiteThread siteThread = new SiteThread(site, this, pageIndexService, siteService);
+                thread = new Thread(siteThread);
+                thread.start();
+            }
         }
     }
 
@@ -47,18 +50,13 @@ public class SiteManagerImpl implements SiteManager {
     }
 
     @Override
-    public void stopParsing() {
-        if (site != null) {
+    public synchronized void stopParsing() {
+        if (!stop) {
             stop = true;
-            site.setStatus(SiteStatus.FAILED);
-            site.setLastError("Индексация отменена");
-            siteService.updateSiteByUrl(site);
-            try {
-                if (thread != null && thread.isAlive()) {
-                    thread.join();
-                }
-            } catch (Exception e) {
-                logger.error(e.toString());
+            if (site != null) {
+                site.setStatus(SiteStatus.FAILED);
+                site.setLastError("Индексация отменена");
+                siteService.updateSiteByUrl(site);
             }
         }
     }
@@ -69,9 +67,18 @@ public class SiteManagerImpl implements SiteManager {
     }
 
     @Override
-    public void deleteSite() {
-        stopParsing();
-        siteService.deleteSiteByUrl(site);
-        site = null;
+    public synchronized void deleteSite() {
+        if (site != null) {
+            stopParsing();
+            try {
+                if (thread != null && thread.isAlive()) {
+                    thread.join();
+                }
+            } catch (Exception e) {
+                logger.error(e.toString());
+            }
+            siteService.deleteSiteByUrl(site);
+            site = null;
+        }
     }
 }
